linux-gen: pool: use 64-bit shm sizes
odp_shm_reserve() takes a 64-bit size argument. The previous 32-bit values
could wrap around in case of very large pools.

Fixes: https://bugs.linaro.org/show_bug.cgi?id=3632

+++ platform/linux-generic/odp_pool.c
@@ -478,8 +478,8 @@ static odp_pool_t pool_create(const char *name, odp_pool_param_t *params,
 	pool->tailroom       = tailroom;
 	pool->block_size     = block_size;
 	pool->uarea_size     = uarea_size;
-	pool->shm_size       = (num + num_extra) * block_size;
-	pool->uarea_shm_size = num * uarea_size;
+	pool->shm_size       = (num + num_extra) * (uint64_t)block_size;
+	pool->uarea_shm_size = num * (uint64_t)uarea_size;
 	pool->ext_desc       = NULL;
 	pool->ext_destroy    = NULL;
 
@@ -956,9 +956,9 @@ void odp_pool_print(odp_pool_t pool_hdl)
 	ODP_PRINT("  tailroom        %u\n", pool->tailroom);
 	ODP_PRINT("  block size      %u\n", pool->block_size);
 	ODP_PRINT("  uarea size      %u\n", pool->uarea_size);
-	ODP_PRINT("  shm size        %u\n", pool->shm_size);
+	ODP_PRINT("  shm size        %" PRIu64 "\n", pool->shm_size);
 	ODP_PRINT("  base addr       %p\n", pool->base_addr);
-	ODP_PRINT("  uarea shm size  %u\n", pool->uarea_shm_size);
+	ODP_PRINT("  uarea shm size  %" PRIu64 "\n", pool->uarea_shm_size);
 	ODP_PRINT("  uarea base addr %p\n", pool->uarea_base_addr);
 	ODP_PRINT("\n");
 }
